Implement keybinding persistence via user command
When using the Inputkb user command, you can choose to persist the keybinding. This will involve writing 'vim.keymap.set(...)' text to a file which your config will then source. This location can be specified in the table passed to the `opts=` in your lazy.nvim config or the `require('keytex').setup()` setup function.